+++ src/azathoth/prompting/couchdb.py
