+++ src/azathoth/prompting/handlers/base_model_handler.py
