+++ src/azathoth/prompting/model_registry.py
