+++ src/azathoth/ui/prompt-ui/public/android-chrome-192x192.png
