+++ src/azathoth/ui/prompt-ui/public/android-chrome-512x512.png
