+++ src/azathoth/ui/prompt-ui/public/apple-touch-icon.png
