+++ src/azathoth/ui/prompt-ui/public/azathoth.png
