+++ src/azathoth/ui/prompt-ui/public/favicon-16x16.png
