+++ src/azathoth/ui/prompt-ui/public/favicon-32x32.png
