+++ src/azathoth/ui/prompt-ui/public/favicon.ico
