+++ src/azathoth/ui/prompt-ui/public/index.html
