+++ src/azathoth/ui/prompt-ui/public/manifest.json
