+++ src/azathoth/ui/prompt-ui/public/site.webmanifest
