+++ src/azathoth/ui/prompt-ui/src/App.css
