+++ src/azathoth/ui/prompt-ui/src/components/DynamicListSimple.js
