+++ src/azathoth/ui/prompt-ui/src/components/PromptGoalEditor.js
