+++ src/azathoth/ui/prompt-ui/src/components/PromptGoalList.js
